Fixed course ownership rights

+++ controllers/courseController.js
@@ -47,11 +47,17 @@ const getCourse = asyncHandler(async (req, res, next) => {
 
 const addCourse = asyncHandler(async (req, res, next) => {
     req.body.bootcamp = req.params.bootcampId;
+    req.body.user = req.user.id;
 
     const bootcamp = await Bootcamp.findById(req.params.bootcampId);
 
     if (!bootcamp) {
         return next(new ErrorResponse(`No bootcamp with the id of ${req.params.bootcampId}`), 404);
+    }
+
+    // Make sure user is bootcamp owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp._id}`, 401));
     }
 
     const course = await Course.create(req.body);
@@ -94,6 +100,11 @@ const deleteCourse = asyncHandler(async (req, res, next) => {
     
         if (!course) {
             return next(new ErrorResponse(`No course with the id of ${req.params.id}`), 404);
+        }
+
+        // Make sure user is course owner
+        if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+            return next(new ErrorResponse(`User ${req.user.id} is not authorized to delete course ${course._id}`, 401));
         }
     
         await course.deleteOne();
